Updated to warn about the history of the file
The information in this file is still useful but we need to remember that our ensembl branch has diverged from anacode's original work. Therefore the information in this doc must be treated with a degree of caution if it refers to now out of date code. The concepts are still correct.